[WIP] Autocomplete brands and models
The database is crowded with bogus entries (see #90). Here's a prototype of autocompletion for watch brands and models. You can try it out by typing 'j'  in the brand, it will suggests 'Java', 'Javascript'  and 'Clojure'  because they all contain a 'j'. If 'Java' is choosen, then a 'n' in the model input will suggests 'Niark'. It also works for 'Scala' -> 'Plop'.

@VS2000, I need supported brands and models of each brand to complete this.