Refactor email management from model to library. #35
Previously, automatic email management was in a model, which make not much sense because we don't have any email table. Also, to be sure that the emails weren't sent twice, we use to store all the computed emails in three different table (email_watch; email_user and email_measure). This wasn't going to be sustainable as the tables grow (where not in hell). Now; we have a table email_batch that stores the last ti,e we computed the emails and we compute between the last time and time() (+/- offset)