fix(login): Mobile Facebook login to website created accounts is now possible
Facebook login always use POST even when the account is alreay created. This is by design because we

don't have the facebook password. An error in string concatenation (+ instead of .) led all facebook

account to the same sha256 password... This password now becomes the default apssword for fb

accounts and a way to distinguish them from classical accounts. In addition, login to or creating

accounts using the reserved fb password is forbidden

Fixes Toolwatchapp/tw-mobile#88